+++ src/dolphin/cli.py
@@ -1,6 +1,7 @@
 import argparse
 import sys
 
+import dolphin._cli_timeseries
 import dolphin._cli_unwrap
 import dolphin.workflows._cli_config
 import dolphin.workflows._cli_run
@@ -20,6 +21,7 @@ def main(args=None):
     dolphin.workflows._cli_run.get_parser(subparser, "run")
     dolphin.workflows._cli_config.get_parser(subparser, "config")
     dolphin._cli_unwrap.get_parser(subparser, "unwrap")
+    dolphin._cli_timeseries.get_parser(subparser, "timeseries")
     parsed_args = parser.parse_args(args=args)
 
     arg_dict = vars(parsed_args)