+++ src/dolphin/workflows/config/_common.py
@@ -124,12 +124,12 @@ class InterferogramNetwork(BaseModel, extra="forbid"):
     max_bandwidth: Optional[int] = Field(
         None,
         description="Max `n` to form the nearest-`n` interferograms by index.",
-        gt=1,
+        ge=1,
     )
     max_temporal_baseline: Optional[int] = Field(
         None,
         description="Maximum temporal baseline of interferograms.",
-        gt=0,
+        ge=0,
     )
     indexes: Optional[list[tuple[int, int]]] = Field(
         None,
@@ -163,6 +163,16 @@ class UnwrapOptions(BaseModel, extra="forbid"):
         description=(
             "Whether to run the unwrapping step after wrapped phase estimation."
         ),
+    )
+    run_goldstein: bool = Field(
+        False,
+        description=(
+            "Whether to run Goldstein filtering step on wrapped interferogram."
+        ),
+    )
+    run_interpolation: bool = Field(
+        False,
+        description=("Whether to run interpolation step on wrapped interferogram."),
     )
     _directory: Path = PrivateAttr(Path("unwrapped"))
     unwrap_method: UnwrapMethod = UnwrapMethod.SNAPHU
@@ -203,6 +213,31 @@ class UnwrapOptions(BaseModel, extra="forbid"):
     cost: Literal["defo", "smooth"] = Field(
         "smooth",
         description="Statistical cost mode method for SNAPHU.",
+    )
+    zero_where_masked: bool = Field(
+        False,
+        description=(
+            "Set wrapped phase/correlation to 0 where mask is 0 before unwrapping. "
+        ),
+    )
+    alpha: float = Field(
+        0.5,
+        description=(
+            "(for Goldstein filtering) Power parameter for Goldstein algorithm."
+        ),
+    )
+    max_radius: int = Field(
+        51,
+        ge=0.0,
+        description=("(for interpolation) maximum radius to find scatterers."),
+    )
+    interpolation_cor_threshold: float = Field(
+        0.5,
+        description=" Threshold on the correlation raster to use for interpolation. "
+        "Pixels with less than this value are replaced by a weighted "
+        "combination of neighboring pixels.",
+        ge=0.0,
+        le=1.0,
     )
 
     @field_validator("ntiles", "downsample_factor", mode="before")